Revert "sys/src/9k/port: use uintptr type for i and l (syscall compatibility with 9front)"
This reverts commit f7a273dbb8a387c837f35c562c4ad7c35d3e3bf1.

+++ sys/src/9k/port/portdat.h
@@ -600,8 +600,8 @@ struct Schedq
 
 typedef union Ar0 Ar0;
 union Ar0 {
-	uintptr	i;
-	uintptr	l;
+	int	i;
+	long	l;
 	uintptr	p;
 	usize	u;
 	void*	v;
